chore(createObserver): remove callback param, move callback fn
