+++ src/__tests__/IntersectionObserver.spec.js
@@ -2,8 +2,8 @@
 import 'intersection-observer';
 import React from 'react';
 import renderer from 'react-test-renderer';
-import IntersectionObserver, { callback } from '../IntersectionObserver';
-import { findObserverElement, observerElementsMap } from '../observer';
+import IntersectionObserver from '../IntersectionObserver';
+import { callback, findObserverElement, observerElementsMap } from '../observer';
 
 function mockUtilsFunctions() {
     const utils = require.requireActual('../utils');
